Move Country choices to separate module
Prevent it from being loaded when importing this library to application.
